streamlit/requirements.txt
streamlit==1.29.0
langchain==0.0.350
langchain-community==0.0.10
langchain-core==0.1.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
plotly==5.18.0
numpy==1.26.2
google-generativeai==0.3.1
google-cloud-aiplatform==1.36.4
chromadb==0.4.18
tiktoken==0.5.2
